ENG-10472: Advance the SPI truncation point when transaction is done.
Advancing the truncation point when processing messages in the
SpScheduler may cause race conditions for MP. When it processes the
CompleteTransactionMessage in the SpScheduler, the transaction may not
be marked done yet, so it's unsafe to advance the truncation
handle. Since there is no response for CompleteTransactionMessage, the
SpScheduler may not know when exactly the CompleteTransactionTask will
mark the transaction done.

The safest place to advance the truncation handle is actually when we
mark transactions done inside the tasks (SpProcedureTask, FragmentTask,
or CompleteTransactionTask). Since this happens on the Site thread, I
synchronized the block that advances the truncation handle in the
SpScheduler.